TestsRosicAndRapt.jucer: set warning level to medium in release build

+++ Tests/TestsRosicAndRapt/Source/rosic_tests/rosic_StringTests.cpp
@@ -58,7 +58,7 @@ void rotes::testStringIntConversions(int numIterations)
   rsString numString;
   for(int i=0; i<numIterations; i++)
   {
-    numberOriginal      = randomUniform(INT_MIN, INT_MAX);
+    numberOriginal      = (int) randomUniform(INT_MIN, INT_MAX);
     numString           = numberOriginal;
     numberReconstructed = numString.asInt();
     rassert( numberReconstructed == numberOriginal );
